Control of possible critical problems.
Control of possible critical problems.

+++ dilogin/src/main/java/di/dilogin/minecraft/cache/UserSessionCache.java
@@ -49,7 +49,7 @@ public class UserSessionCache {
 		if (!user.getIp().equals(ip))
 			return false;
 
-		if (Calendar.getInstance().getTimeInMillis() > ((Long) sessions.get(user)).longValue())
+		if (Calendar.getInstance().getTimeInMillis() > (sessions.get(user)).longValue())
 			return false;
 		
 		if (!userDao.contains(name)) {
